requirements.txt
flask>=2.0.0
requests>=2.25.0
python-dotenv>=0.19.0
transformers>=4.30.0
torch>=2.0.0
gunicorn==21.2.0 